Add files via upload
Imagine you are working as a data scientist for a manufacturing company. They are using vibration sensors to monitor the machinery within their production line and want to use this data for predictive maintenance – the aim is to be able to classify whether there is a fault with the machine based on the readings from the vibration sensors. We have provided you with a dataset FaultDataset.csv. Each row contains twenty vibration sensor readings, and the final column identifies whether there was a fault with the machine at the time of the readings. In this column, 0 means there was no fault with the machine and 1 means a fault was identified.